text_email_detection/requirements.txt
torch
transformers
tokenizers
regex
pdfplumber
python-docx
email-validator
numpy
tqdm
